Update: path 'solution/solution.cc' - Updated by 'notice99'

+++ solution/solution.cc
@@ -1,7 +1,7 @@
 #include <bits/stdc++.h>
 using namespace std;
 int main() {
-    int a;
+    long long a;
     cin >> a;
     cout << a*(a+1)/2;
     return 0;
